requirements.txt
uvicorn==0.34.0
fastapi==0.115.6
jose==1.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.12.0
PyJWT==2.10.1
SQLAlchemy==2.0.41
requests==2.32.3
python-multipart==0.0.20
alembic==1.17.2
asyncpg==0.31.0
psycopg2==2.9.11
python-dotenv==1.2.1